[TM-1736] Upgrade to NestJS 11
