+++ apps/unified-database-service/src/airtable/airtable.module.ts
@@ -5,7 +5,8 @@ import { BullModule } from "@nestjs/bullmq";
 import { Module } from "@nestjs/common";
 import { AirtableService } from "./airtable.service";
 import { AirtableProcessor } from "./airtable.processor";
-import { QueueHealthService } from "./queue-health.service";
+import { QueueHealthIndicator } from "./queue-health.indicator";
+import { TerminusModule } from "@nestjs/terminus";
 
 @Module({
   imports: [
@@ -25,9 +26,10 @@ import { QueueHealthService } from "./queue-health.service";
         }
       })
     }),
-    BullModule.registerQueue({ name: "airtable" })
+    BullModule.registerQueue({ name: "airtable" }),
+    TerminusModule
   ],
-  providers: [AirtableService, AirtableProcessor, QueueHealthService],
-  exports: [AirtableService, QueueHealthService]
+  providers: [AirtableService, AirtableProcessor, QueueHealthIndicator],
+  exports: [AirtableService, QueueHealthIndicator]
 })
 export class AirtableModule {}
