+++ apps/unified-database-service/src/health/health.controller.ts
@@ -3,7 +3,7 @@ import { HealthCheck, HealthCheckService, SequelizeHealthIndicator } from "@nest
 import { NoBearerAuth } from "@terramatch-microservices/common/guards";
 import { ApiExcludeController } from "@nestjs/swagger";
 import { User } from "@terramatch-microservices/database/entities";
-import { QueueHealthService } from "../airtable/queue-health.service";
+import { QueueHealthIndicator } from "../airtable/queue-health.indicator";
 
 @Controller("health")
 @ApiExcludeController()
@@ -11,7 +11,7 @@ export class HealthController {
   constructor(
     private readonly health: HealthCheckService,
     private readonly db: SequelizeHealthIndicator,
-    private readonly queue: QueueHealthService
+    private readonly queue: QueueHealthIndicator
   ) {}
 
   @Get()
@@ -20,10 +20,7 @@ export class HealthController {
   async check() {
     const connection = await User.sequelize.connectionManager.getConnection({ type: "read" });
     try {
-      return this.health.check([
-        () => this.db.pingCheck("database", { connection }),
-        () => this.queue.queueHealthCheck()
-      ]);
+      return this.health.check([() => this.db.pingCheck("database", { connection }), () => this.queue.isHealthy()]);
     } finally {
       User.sequelize.connectionManager.releaseConnection(connection);
     }
