+++ package.json
@@ -14,18 +14,18 @@
   "private": true,
   "dependencies": {
     "@casl/ability": "^6.7.3",
-    "@nestjs/bullmq": "^10.2.3",
-    "@nestjs/common": "^10.0.2",
-    "@nestjs/config": "^3.2.3",
-    "@nestjs/core": "^10.0.2",
-    "@nestjs/jwt": "^10.2.0",
-    "@nestjs/platform-express": "^10.0.2",
-    "@nestjs/schedule": "^4.1.2",
-    "@nestjs/sequelize": "^10.0.1",
-    "@nestjs/swagger": "^8.1.0",
-    "@nestjs/terminus": "^10.2.3",
-    "@sentry/nestjs": "8.54.0",
-    "@sentry/profiling-node": "^8.54.0",
+    "@nestjs/bullmq": "^11.0.2",
+    "@nestjs/common": "^11.0.11",
+    "@nestjs/config": "^4.0.1",
+    "@nestjs/core": "^11.0.11",
+    "@nestjs/jwt": "^11.0.0",
+    "@nestjs/platform-express": "^11.0.11",
+    "@nestjs/schedule": "^5.0.1",
+    "@nestjs/sequelize": "^11.0.0",
+    "@nestjs/swagger": "^11.0.6",
+    "@nestjs/terminus": "^11.0.0",
+    "@sentry/nestjs": "9.6.0",
+    "@sentry/profiling-node": "^9.6.0",
     "@slack/web-api": "^7.8.0",
     "@transifex/native": "^7.1.3",
     "airtable": "^0.12.2",
@@ -39,7 +39,7 @@
     "luxon": "^3.5.0",
     "mariadb": "^3.4.0",
     "mysql2": "^3.13.0",
-    "nestjs-request-context": "^3.0.0",
+    "nestjs-request-context": "^4.0.0",
     "nodemailer": "^6.10.0",
     "reflect-metadata": "^0.2.2",
     "rxjs": "^7.8.2",
@@ -50,8 +50,8 @@
   "devDependencies": {
     "@faker-js/faker": "^9.6.0",
     "@golevelup/ts-jest": "^0.6.2",
-    "@nestjs/schematics": "^10.0.1",
-    "@nestjs/testing": "^10.0.2",
+    "@nestjs/schematics": "^11.0.2",
+    "@nestjs/testing": "^11.0.11",
     "@nx/eslint": "20.6.1",
     "@nx/eslint-plugin": "20.6.1",
     "@nx/jest": "20.6.1",
